Add integration test wip

+++ src/main/scala/pl/bondyra/smaz/input/Input.scala
@@ -2,5 +2,5 @@ package pl.bondyra.smaz.input
 
 trait Input {
   def identifier: String
-  def eventTime: Double
+  def eventTime: Long
 }
